Move ConsoleReporter to own file.
