+++ src/reporter/mod.rs
@@ -1,3 +1,4 @@
 pub mod base;
 pub mod carbon;
+pub mod console;
 pub mod prometheus;
